et-grpc-server/requirements.txt
beautifulsoup4==4.12.3
black==24.4.2
cachetools==4.2.4
cassandra-driver==3.25.0
certifi==2024.6.2
chardet==3.0.4
click==8.1.7
geomet==0.2.1.post1
google==2.0.3
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.0
grpcio==1.38.0
grpcio-tools==1.27.2
httplib2==0.22.0
idna==2.10
isort==5.13.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
protobuf==3.14.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.9
six==1.16.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==1.25.11
